Move `SubgraphIndex` implementation to `subgraph` module
